code cleanup

+++ Sailing/ActivityItem.cs
@@ -12,6 +12,6 @@ namespace Sailing
         public string ActivityDesc { get; set; }
         public string ActivityType { get; set; }
         public string ActivityLocation { get; set; }
-        public float EntranceFee { get; set; }
+        public double EntranceFee { get; set; }
     }
 }
